pyproject.toml
[project]
name = "job-recommender-system-with-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "apify-client>=2.1.0",
    "openai>=1.109.1",
    "pymupdf>=1.26.4",
    "python-dotenv>=1.1.1",
    "streamlit>=1.50.0",
]
